refactor(core): move essential words out of interpreter's file
Don't want to mix the behaviors of the VM with its structure.
